Python/Batch/requirements.txt
azure-batch==8.0.0
azure-storage-blob==1.3.1
